Update SwappingContract.sol
Fee Mechanism:

The contract retains a percentage of the swapped tokens as a fee, which can be adjusted by the owner using setFeePercent.
The fee is deducted from the total amount of tokens before transferring them to the user.
Price Slippage Protection:

A slippage parameter is included in the swapEtherToToken function to allow users to specify their tolerance for slippage. If the actual token amount received is less than the acceptable slippage, the transaction is reverted.
ERC20 Token Approval Handling:

The contract supports token-to-token swaps. Users can now swap one ERC20 token for another by calling swapTokenToToken. The contract requires the user to first approve the transfer of tokens via transferFrom.